refactor folder and code
