+++ HotTipster/HotTipster/BusinessLogic/RaceCourseEnum.cs
