+++ HotTipster/HotTipster/GUI/HotTipsterForm.cs
